+++ src/main/java/com/dobugs/yologaauthenticationapi/support/kakao/KakaoConnector.java
@@ -1,7 +1,10 @@
-package com.dobugs.yologaauthenticationapi.support;
+package com.dobugs.yologaauthenticationapi.support.kakao;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
+
+import com.dobugs.yologaauthenticationapi.support.OAuthConnector;
+import com.dobugs.yologaauthenticationapi.support.OAuthProvider;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
